==> notebooks/Supplementary_figures/Suppl_fig8/panel_A/process_external_spatial_datasets/itzhak2016stcSILAC-NOC/fraction/4.NOC_processing.ipynb <==
# NOC processing

import os
import sys
from pathlib import Path
import pandas as pd

script_path = Path.cwd().parent.parent.parent.parent.parent.parent.parent / "script"
data_path = Path.cwd().parent.parent.parent.parent.parent.parent.parent / "data"
sys.path.append(str(script_path))

### Load NOC table


#%store -r timestamp
timestamp = "2024-07-16"

print(f"Timestamp: {timestamp}")

outprefix = f"{timestamp}_QC_filter_impute"

outdir = Path.cwd() / "output"
preprocessing_out = outdir / f"preprocessing"
NOC_path = data_path / "external" / f"elife-16950-supp9-v3.xlsx"

try:
    cyt_table = pd.read_excel(NOC_path, sheet_name="LFQ Static Cyt")
    nuc_table = pd.read_excel(NOC_path, sheet_name="LFQ Static Nuc")
    org_table = pd.read_excel(NOC_path, sheet_name="LFQ Static Org")
except FileNotFoundError:
    print(f"File {NOC_path} not found.\nPlease rerun the previous steps or specify the correct timestamp, current value is {timestamp}")
except pd.errors.ParserError:
    print(f"There was an error parsing the CSV file at {NOC_path}.")
except Exception as e:
    print(f"An unexpected error occurred: {e}")

### Load fraction table

enrich_out_dir = outdir / "fraction_tables"
os.makedirs(enrich_out_dir, exist_ok=True)

fraction_csv_path = enrich_out_dir / f"{timestamp}_fraction_table.csv"

final_fraction_table = pd.read_csv(fraction_csv_path, header=[0, 1], index_col=0)


### Calculation of proportion

# compute median of 6 MAPS
cyt_table["median LFQ intensity cytosolic"] = cyt_table.iloc[:, 4:10].median(axis=1)
nuc_table["median LFQ intensity nuclear"] = nuc_table.iloc[:, 4:10].median(axis=1)
org_table["median LFQ intensity organelle"] = org_table.iloc[:, 4:10].median(axis=1)

def compute_proportion(row):
    uniprot_id = row[("metadata", "Unnamed: 0")]
    if not uniprot_id in list(cyt_table["Lead IDs"]):
        cyt_intensity = 0
    else:
        cyt_intensity = cyt_table.loc[cyt_table["Lead IDs"] == uniprot_id, "median LFQ intensity cytosolic"].values[0]
    if not uniprot_id in list(nuc_table["Lead ID"]):
        nuc_intensity = 0
    else:
        nuc_intensity = nuc_table.loc[nuc_table["Lead ID"] == uniprot_id, "median LFQ intensity nuclear"].values[0]
    if not uniprot_id in list(org_table["Lead ID"]):
        org_intensity = 0
    else:
        org_intensity = org_table.loc[org_table["Lead ID"] == uniprot_id, "median LFQ intensity organelle"].values[0]
    total_intensity = cyt_intensity + nuc_intensity + org_intensity
    return nuc_intensity/total_intensity, org_intensity/total_intensity, cyt_intensity/total_intensity

final_fraction_table[[("sample","NOC_nuclear"), ("sample","NOC_organelle"), ("sample","NOC_cytosol")]] = final_fraction_table.apply(lambda row: pd.Series(compute_proportion(row)), axis=1)

# rename
final_fraction_table = final_fraction_table.rename(columns={("metadata","Unnamed: 0"): ("metadata","uniprot_id")})

final_fraction_table.columns = pd.MultiIndex.from_tuples(
    [(upper, lower.replace('Unnamed: 0', 'Protein IDs')) for upper, lower in final_fraction_table.columns]
)

# remove proteins where the sum of all NOC fractions is 0
final_fraction_table = final_fraction_table[final_fraction_table[("sample","NOC_nuclear")] + final_fraction_table[("sample","NOC_organelle")] + final_fraction_table[("sample","NOC_cytosol")] > 0]

final_fraction_table

### save to file

# saving the final tables to files
enrich_out_dir = outdir / "fraction_tables"
os.makedirs(enrich_out_dir, exist_ok=True)

save_csv_path = enrich_out_dir / f"{timestamp}_fraction_table_NOC.csv"

final_fraction_table.to_csv(save_csv_path)
